vsock: Handle the single VM case same as multiple VMs
In future, we could add the ability to change the configuration at runtime
and allow new guests to be added even without having to restart the
daemon. So it is reasonable to not differentiate between the single and
multiple VM cases, and always pass the `cid_map` to the
VhostUserVsockBackend, even with only one guest.

Signed-off-by: Priyansh Rathi <techiepriyansh@gmail.com>