vsock: Add new VMs at runtime
Implement ability to add new VMs at runtime. Can also modify the
configuration for existing VMs but the modifications will take effect only
after the VM is restarted. Removing a VM at runtime is not supported at
the moment.

To use this feature, you must use the `--config` CLI argument to specify
the initial configuration with the `--watch` flag set. The path to the
config file provided must be a symlink to the actual YAML config. To
update the VM configuration at runtime, simply change the symlink to point
to the YAML config file with the new configuration.

Signed-off-by: Priyansh Rathi <techiepriyansh@gmail.com>